Moved is_ts_graph to _11stereo
